Massive request update
Factions support (beta):
https://www.spigotmc.org/resources/koth2factions.11635/

Scoreboard changes:
The plugin is now using the main scoreboard. This means that there
should be a lot more support with other scoreboard plugins!

Command changes:
The "/koth" command is now COMPLETELY configurable. people that do not
have permission to the admin permission will see it. This is changable
in the config. (as admin, you can look at it by doing "/koth asmember")

There is now a "/koth reload"! I tested it quite a lot, and everything
seems to be working. Might still have some little things I missed. If
so, please mention them here.

Config changes:
- A completely new layout for the config which should be easier to
change/use.
- Added: "remove-lootchest-after", which gives you the ability to
automatically remove the chest after X seconds.
- Added: "drop-loot-on-removal", which like it says, will drop the loot
when the chest gets removed.
- Added: "minimum-players-needed", so a KoTH doesn't start with only 2
people on the server (is configurable ofcourse) (Only works on schedule
(doesn't block starts on command))
- Added: "instant-loot", instead of spawning a chest, it will
automatically give it to the player. (Helpful for crate keys and such!).
Note: For now, if there is no space in the players inventory it will
drop on the floor instead!

Lang.json updates:
- Added: one for if the player doesn't have enough space in their
inventory: (Koth_won_dropping_items)
- Added: an entry for if the player doesn't have permission to a certain
command (this is kind of only used if the player doesn't have permission
to any command) (command_no_permission)
- Added: an entry for when the reload completed (command_reload)

Development changes:
- Changed: KothLeftEvent is now cancellable (#Factions)
- Changed: KothDummy is now updating when it gets changed. It now also
has an isInArea(OfflinePlayer) function. Thought that might be helpful