Fix for non-faction member trying to cap
Apparently you cannot catch an error thrown if you try to initialize it
through reflection.
So the point of NoCompatibleCapperException is useless and I moved the
checking to the getCapper function. This is a sad moment in my life ;(

This fix will stop the exception showing up in console if a non-faction
member tries to cap the KoTH